Implement default mouse effect logic
Default mouse effect logic is on mouseover, the given pixel will change
it's background-color to black.

Side-Effects of this implementation is that I'm not sure if keeping this
implementation will be possible when adding in RGB & Shade mouse effects
later down the line. Probably will need a rework to be adaptable.