Implement grid options logic for settings
Like with setPixelListeners(), setSettingsListeners() utilizes event
delegation to avoid unnecessary bloat of event listeners on the page.

Side-effects of this is that setSettingsListeners() will contain the
sub-logic for each of the individual options / settings.

Unintuitive consequence of clearGrid() current implementation is that
you CANNOT swap the reset of the background color logic with this:
 - pixel.setAttribute("style", "background-color: white;")
Because, that will destroy all of the other CSS styles applied to the
pixel, when the pixels are initialized or updated with setGridDensity().

This wasn't caught earlier with setPixelListeners() due to pure luck.

This means that for any mouse effects implementation logic, you CANNOT
use something like .setAttribute("style", ...) mindlessly.