Fix clearGrid to work on mouse shade effect
Before, when one used the mouse shade effect, the pixels that were
shaded via brightness, were not reset in their brightness w/ clearGrid.
This is an issue, because clearGrid is meant to reset pixels to default
values without adjusting grid density.

Now, the brightness is reset to default value when clearGrid is ran.