Implement temporary pixel border inking
Before, the black border inking about each pixel in the sketch pad
would never disappear, which was great during development, but not for
actual usage, especially as the pixel density got higher, which led to
the border taking up substantial sketch-pad real estate.

Now, on page load, and after the grid density is adjusted, the black
border inking will disappear after 3 seconds - just enough time to
visually confirm what the current and new grid densities are, but then
it disappears so the sketch-pad drawing can be cleaner.