Fix sketchpad so its background does not change
Before, the <div> with class sketch-pad, would change it's background
color, if you hovered the mouse in the negative space of the corner
pixels with the default effect or rgb effect, but not shade effect.
What was wrong with this, is that that class's background is never to
change in the first place to any event.

Now, as intended, that class's background doesn't change from events.

The logic behind the bug, was due to the sketch-pad div being used as an
event delegator. Normally it wouldn't ever be the target itself, as
each of the pixels would take up the entirety of its hit-box. But due to
the corner pixels being rounded for aesthetic reasons, it led to the
sketch-pad div hit-box being available.

And why it didn't change background on shade effect, is because the
sketch-pad div class didn't have any filter style to begin with, which
is how the shade effect works.

Why I chose the current approach is knowing that sketch-pad doesn't have
an id, unlike the rest of the pixels it contains. A side-effect of this
is that if I ever wanted to add an id to it, I'd need to rework the
implementation of this bug fix to check for targets with ids that are
strictly numerical without any parsing instead.